refactor(types): split header into v1 and v2 submodules
